Implemented action selection and action implementation in ai system

+++ DrifterEngine/src/Utility/EntityHelpers.h
@@ -6,5 +6,5 @@ namespace drft::util
 	RenderComponent getRenderData(entt::const_handle entity);
 	std::string getEntityName(const entt::const_handle entity);
 	bool isType(entt::const_handle entity, const std::string& typeName);
-	int getDistanceBetween(entt::const_handle e1, entt::const_handle e2);
+	float getDistanceBetween(entt::const_handle e1, entt::const_handle e2);
 }
